YT-7773: Generate less comparers

+++ yt/ytlib/query_client/evaluation_helpers.h
@@ -48,7 +48,7 @@ static const size_t InitialGroupOpHashtableCapacity = 1024;
 
 using THasherFunction = ui64(const TValue*);
 using TComparerFunction = char(const TValue*, const TValue*);
-using TTernaryComparerFunction = int(const TValue*, const TValue*);
+using TTernaryComparerFunction = i64(const TValue*, const TValue*);
 
 namespace NDetail {
 class TGroupHasher
